+++ CHANGELOG.md
@@ -8,8 +8,6 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 ### Added
 
 ### Changed
-- Change default language.
-- Add zh-Hant localization.
 
 ### Deprecated
 
@@ -18,6 +16,13 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 ### Fixed
 
 ### Security
+
+## [1.3.0] - 2026-06-03
+### Added
+- Add zh-Hant localization.
+
+### Changed
+- Change default language.
 
 ## [1.2.1] - 2025-10-02
 ### Fixed